windows/desktop/requirements.txt
pyinstaller>=6.14.1,<7
pywebview>=5.4,<6